+++ Example/HttpInterceptor/Template/ContentTypeEnum.swift
