+++ Example/HttpInterceptor/Template/MockerViewController.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+class MockerViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Mocker"
+
+    }
+
+
+}
