+++ Example/HttpInterceptor/Template/SessionManager.swift
