+++ HttpInterceptor/Classes/HttpMockerDelegate.swift
@@ -1,0 +1,1 @@
+import Foundation
